type_as -> to

+++ retnet/modeling_retnet.py
@@ -511,7 +511,7 @@ class MultiScaleRetention(nn.Module):
         dtype = retention_out.dtype
         # when elementwise_affine=False, apex.normalization.FusedRMSNorm may autocast to
         # fp32. We want it back to original dtype.
-        normed = self.group_norm(retention_out).reshape(B, T, self.value_dim).type_as(dtype)
+        normed = self.group_norm(retention_out).reshape(B, T, self.value_dim).to(dtype)
         # out gate & proj
         out = self.gate_fn(g) * normed
         out = self.out_proj(out)
